Add 'working' un-indent.
